Bugfixes for cocreate and cosend

+++ include/comsg/coport_ipc.h
@@ -35,12 +35,13 @@
 #include <coproc/coport.h>
 
 #include <stddef.h>
+#include <sys/types.h>
 
 nsobject_t *open_named_coport(const char *coport_name, coport_type_t type, namespace_t *ns);
 coport_t *open_coport(coport_type_t type);
 
-int cosend(const coport_t *prt, const void *buf, size_t len);
-int corecv(const coport_t *port, void **buf, size_t len);
+ssize_t cosend(const coport_t *prt, const void *buf, size_t len);
+ssize_t corecv(const coport_t *port, void **buf, size_t len);
 coport_type_t coport_gettype(coport_t *port);
 void make_pollcoport(pollcoport_t *pct, coport_t *port, coport_eventmask_t events);
